Remove/Clean some containers and add configureStore
