+++ src/app/actions/Counter.js
@@ -1,5 +1,4 @@
-export const INCREMENT_COUNTER = 'INCREMENT_COUNTER';
-export const DECREMENT_COUNTER = 'DECREMENT_COUNTER';
+import { INCREMENT_COUNTER, DECREMENT_COUNTER } from '../constants/ActionTypes';
 
 export function increment() {
   return {
@@ -16,7 +15,6 @@ export function decrement() {
 export function incrementIfOdd() {
   return (dispatch, getState) => {
     const { counter } = getState();
-
     if (counter % 2 === 0) {
       return;
     }
