--- src/app/components/Main.js
@@ -1,16 +1,0 @@
-import React, { Component, PropTypes } from 'react';
-
-export default class Main extends Component {
-
-  static propTypes = {
-    children: PropTypes.any.isRequired
-  }
-
-  render() {
-    return (
-      <div>
-          {this.props.children}
-      </div>
-    );
-  }
-}
