+++ src/app/containers/CounterApp.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import { bindActionCreators } from 'redux';
+import { connect } from 'react-redux';
+import Counter from '../components/Counter';
+import * as CounterActions from '../actions/Counter';
+
+@connect(state => ({
+  counter: state.counter
+}))
+export default class CounterApp {
+  render() {
+    const { counter, dispatch } = this.props;
+    return (
+      <Counter counter={counter}
+               {...bindActionCreators(CounterActions, dispatch)} />
+    );
+  }
+}
