--- src/app/containers/index.js
@@ -1,3 +1,0 @@
-export { default as Root } from './Root';
-export { default as App } from './App';
-export { default as CounterPage } from './CounterPage';
