+++ src/app/reducers/counter.js
@@ -1,7 +1,9 @@
-import { INCREMENT_COUNTER, DECREMENT_COUNTER } from '../actions/counter';
+import { SET_COUNTER, INCREMENT_COUNTER, DECREMENT_COUNTER } from '../actions/counter';
 
 export default function counter(state = 0, action) {
   switch (action.type) {
+  case SET_COUNTER:
+    return action.payload;
   case INCREMENT_COUNTER:
     return state + 1;
   case DECREMENT_COUNTER:
